% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{process_observation_logs}
\alias{process_observation_logs}
\title{Processes Dipol-UF and Dipol-2 observation logs
Generates observation databases as plaintext, json, csv and sqlite.}
\usage{
process_observation_logs(path = fs::path("."), out = fs::path(path, "output"))
}
\arguments{
\item{path}{Root directory. `path/"input"` is searched for input files.}

\item{out}{Where to write output.}
}
\description{
Processes Dipol-UF and Dipol-2 observation logs
Generates observation databases as plaintext, json, csv and sqlite.
}
